requirements.txt
coverage

# last know working 5.4.3
pytest
watchdog
python-socketio>=5
aiohttp
PyYAML
pytest-asyncio
nest-asyncio
bump-my-version
